+++ asgi_lua/__init__.py
@@ -6,6 +6,7 @@ from .bundle import pack_bundle, test_bundle, validate_bundle
 from .middleware import LuaConfig, LuaMiddleware
 from .presets import copy_builtin_preset, list_builtin_presets
 from .promotion import compare_decisions, diff_policies
+from .proxy import ReverseProxyApp, create_proxy_application, run_proxy
 from .recorder import append_record, load_record_log, record_audit_event, record_policy_request, replay_record_log
 from .redaction import RedactionConfig, append_audit_event, build_audit_event, redact_secrets
 from .runtime import LuaDecisionError, LuaDecisionTrace, LuaRuntimeError
@@ -29,6 +30,7 @@ __all__ = [
     "MemoryStateStore",
     "RedisStateStore",
     "RedactionConfig",
+    "ReverseProxyApp",
     "SQLiteStateStore",
     "SnapshotStateStore",
     "StateLimits",
@@ -38,6 +40,7 @@ __all__ = [
     "build_audit_event",
     "compare_decisions",
     "copy_builtin_preset",
+    "create_proxy_application",
     "diff_policies",
     "load_record_log",
     "list_builtin_presets",
@@ -46,6 +49,7 @@ __all__ = [
     "record_policy_request",
     "redact_secrets",
     "replay_record_log",
+    "run_proxy",
     "simulate_policy",
     "test_bundle",
     "validate_bundle",